chore: Refactor index, count and set membership
Our implementation of the index, count and __contains__ methods so far were using the standard libreary implementations, which rely on linear search. We now refactored and overridden these implementations in the SortedFrozenSet to use the binary search, and take advantage of the fact, that the SortedFrozenSet is an ordered collection of the distinct elements.